Address review feedback on the perf-optimization branch
- Split the WatchEvent/ScrobbleSession index migration into six single-
  statement migrations using CREATE/DROP INDEX CONCURRENTLY, so the builds
  don't take an exclusive lock on tables that are written to on every
  scrobble/webhook event. Prisma only skips wrapping a migration in a
  transaction when the file has exactly one statement (CONCURRENTLY can't
  run inside one), hence one index operation per file instead of a single
  combined migration. Verified against a fresh Postgres instance.
- Bound the batch watch-event update flush to 10 concurrent requests
  instead of firing them all via one unbounded Promise.all; extracted the
  bounded-concurrency helper import.ts already had into a shared
  lib/concurrency.ts used by both.
- Key the show-details cache by TMDB language in addition to tmdbId (it's
  a per-user setting), and stop caching null results from a fetch error —
  those aren't the same as a legitimate "no next episode" answer, and
  caching them for hours made a transient TMDB outage look like it lasted
  that long.
- Guard the ListsPage infinite-scroll sentinel so it isn't mounted while
  the loading skeleton is showing, and give the IntersectionObserver an
  800px rootMargin so more items render before the user scrolls into the
  empty edge.
